refactor: rename events.py back to handlers.py for clearer diff review
